backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
aiomysql==0.2.0
numpy==1.26.4
python-dotenv==1.0.1
aiofiles==23.2.1
